% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifna.R
\name{ifna}
\alias{ifna}
\title{Replace a missing value (NA) with an alternative}
\usage{
ifna(value, alternative)
}
\arguments{
\item{value}{The missing value to be replaced.}

\item{alternative}{The alternative value to replace.}
}
\description{
This function replaces a missing value (NA) with an alternative
}
\examples{
output <- ifna(value, alternative)
}
