% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_str.R
\name{in_str}
\alias{in_str}
\title{Identify the nth character of a certain string}
\usage{
in_str(x, y)
}
\description{
This function is a modified version of the 'str_replace' function from 'stringr' package.
}
