% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_str.R
\name{left_str}
\alias{left_str}
\title{Replace certain first left-most characters of a string}
\usage{
left_str(x, y)
}
\arguments{
\item{value}{The value of interest.}

\item{int}{An integer that specifies the number of character.}
}
\description{
This function replaces certain first left-most characters of a string
}
