% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_str.R
\name{remove_str}
\alias{remove_str}
\title{Remove a certain value}
\usage{
remove_str(string, pattern)
}
\description{
This function is a straight copy of the 'str_remove_all' function from 'stringr' package.
}
