% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_str.R
\name{replace_str}
\alias{replace_str}
\title{Replace a certain value with an alternative}
\usage{
replace_str(string, pattern, replacement)
}
\description{
This function is a straight copy of the 'str_replace' function from 'stringr' package.
}
