% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_str.R
\name{right_str}
\alias{right_str}
\title{Replace certain first right-most characters of a string}
\usage{
right_str(value, int)
}
\arguments{
\item{value}{The value of interest.}

\item{int}{An integer that specifies the number of character.}
}
\description{
This function replaces certain first right-most characters of a string
}
